Fix never-ending `conda render` bug by forcing `conda-build` v3.20.3
`conda-build` v3.20.4 introduces a bug which causes never-ending
rendering after issuing `conda render`. The only output is:
```
write() argument must be str, not bytes
```
which is printed multiple times.

Based on `litex-conda-eda` testing it affects most of the recipes (e.g.
`vtr`), but not all. `verilator` and `prjtrellis` are the ones from the
"No dependencies" stage that doesn't seem to be affected.