symbiflow-yosys: Modify for macOS
Building on Linux stays exactly the same as in the previous commit.

Single patch modifying `Makefile` and `tests/*` was updated and divided
into two separate patches.

Simple testing with commands gets moved to the `meta.yaml` as there was
a problem with running `yosys -V` because `libffi` couldn't be located
properly via `rpath`.