Set proper `yosys` in packages requiring it
The stage of some packages has changed because `symbiflow-yosys` is
built at the "Has first level dependencies" stage.

Comments were added with required packages built in this repository to
make the stage selection easier in case of any further changes.

`nextpnr-*` packages on Windows are allowed to fail because currently
`symbiflow-yosys` isn't built for Windows.