Install anaconda-client in the base environment
Changes to use `conda-build-prepare` were never tested on the `master`
branch of `litex-conda-*` repositories so the problem with moving
uploaded packages to the 'main` label never occured.

The `anaconda-client` package was installed by the `conda-build-prepare`
which obviously is only called for the jobs that build packages.

Now the missing package will be installed in the `base` environment so
that it can be used by all scripts calling `anaconda`.