Add periodic builds
The packages are usually built from their default branches (the version
is not frozen).

For now we only had builds triggered by changes in the repository with
the packages.

The tool repositories change independently of the conda package
repository.

Adding a periodic build allows us to keep the packages up to date
without the need to update this repo and without triggering anything
manually.

Signed-off-by: Tomasz Gorochowik <tgorochowik@antmicro.com>