script.sh: Exit instantly if `conda-build` fails
Because of piping the output to `tee`, the failing `conda build` command
never caused script to exit. This was highly frustrating for checking
errors in GH Actions UI. The package not found for the installation was
always showed as the reason for the failing job and the actual error had
to be found manually.

Now the building error will be shown straight away.

Includes also a minor fix for the install section naming.