sim/verilator: linux: Require compiler package v8 or newer
Verilator hardcodes the compiler name in the `verilated.mk` file so the
same is then required for verilation. The packages prior to v8 contain
only the `x86_64-conda_cos6-linux-gnu`-prefixed executables.

Currently there are both `...-conda-...` and `...-conda_cosX-...` with
the `cosX` ones being links.

The constraint for macOS is lifted because it turns out it can be
correctly built without it.