yosys: add -Wno-c++11-narrowing to Makefile
There are several bugs in `abc` that cause it to cast `unsigned` to
`int`. The Conda version of clang doesn't like this, and generates
an error.

The default yosys build for clang adds this argument, however the
conda-eda version of the Makefile was missing it.

This fixes builds on arm64.

Signed-off-by: Sean Cross <sean@xobs.io>