openroad: adjust dependencies for macos
The earliest version of spdlog that Conda offers for osx-arm64 is
1.9.0. They also offer 1.9.2 and 1.11.0, but openroad is not compatible
with fmt when using spdlog 1.11.

Pin the version of spdlog to 1.9 to ensure that it builds.

Additionally, libgomp doesn't exist on osx, but llvm provides a
compatible library called `llvm-openmp`. Use that package on osx instead.

Tested on osx-arm64.

Signed-off-by: Sean Cross <sean@xobs.io>